+++ src/test/viewport/index.js
@@ -1,1 +1,0 @@
-import './viewport.spec';