+++ src/viewport/index.js
@@ -1,1 +1,0 @@
-export {default as WebGLViewport} from './viewport';